* Pulling control flow analysis and annotating out out PrimitiveArrayAnalyzer and into StandardControlFlowAnalyzer * Started added generic control flow interfaces
