+++ Phosphor/src/edu/columbia/cs/psl/phosphor/control/ControlFlowPropagationPolicy.java
@@ -1,0 +1,7 @@
+package edu.columbia.cs.psl.phosphor.control;
+
+import edu.columbia.cs.psl.phosphor.PhosphorInstructionInfo;
+
+interface ControlFlowPropagationPolicy {
+    void visitingPhosphorInstructionInfo(PhosphorInstructionInfo info);
+}
